Moved settings related stuff into its own package
