+++ app/src/main/java/de/voicegym/voicegym/menu/NavigationDrawerActivity.kt
@@ -21,6 +21,7 @@ import de.voicegym.voicegym.R
 import de.voicegym.voicegym.recordActivity.RecordActivity
 import de.voicegym.voicegym.util.SwitchToRecordingViewListener
 import de.voicegym.voicegym.menu.dummy.ExerciseContent
+import de.voicegym.voicegym.menu.settings.SettingsActivity
 import de.voicegym.voicegym.model.Recording
 import de.voicegym.voicegym.recordings.RecordingsFragment
 import kotlinx.android.synthetic.main.activity_navigation_drawer.drawer_layout
