+++ app/src/main/java/de/voicegym/voicegym/menu/settings/SettingsActivity.kt
@@ -1,4 +1,4 @@
-package de.voicegym.voicegym.menu
+package de.voicegym.voicegym.menu.settings
 
 import android.annotation.TargetApi
 import android.content.Context
@@ -15,6 +15,7 @@ import android.preference.PreferenceFragment
 import android.preference.PreferenceManager
 import android.view.MenuItem
 import de.voicegym.voicegym.R
+import de.voicegym.voicegym.menu.AppCompatPreferenceActivity
 
 /**
  * A [PreferenceActivity] that presents a set of application settings. On
