+++ app/src/main/java/de/voicegym/voicegym/util/audio/RecordHelper.kt
@@ -1,6 +1,5 @@
 package de.voicegym.voicegym.util.audio
 
-import android.media.AudioFormat.CHANNEL_IN_MONO
 import android.media.AudioFormat.ENCODING_PCM_16BIT
 import android.media.AudioRecord
 import android.media.AudioRecord.STATE_INITIALIZED
@@ -9,9 +8,9 @@ import android.media.MediaRecorder.AudioSource.MIC
 import android.os.Process.THREAD_PRIORITY_AUDIO
 import android.os.Process.setThreadPriority
 import android.util.Log
-import de.voicegym.voicegym.menu.SettingBundle.audioFormat
-import de.voicegym.voicegym.menu.SettingBundle.channelConfig
-import de.voicegym.voicegym.menu.SettingBundle.sampleRate
+import de.voicegym.voicegym.menu.settings.SettingBundle.audioFormat
+import de.voicegym.voicegym.menu.settings.SettingBundle.channelConfig
+import de.voicegym.voicegym.menu.settings.SettingBundle.sampleRate
 import de.voicegym.voicegym.util.RecordBufferListener
 
 class RecordHelper(private val preferredBufferSize: Int) {
